pyproject.toml
[project]
name = "sentinel"
version = "0.2.0"
description = "Defence-in-depth AI assistant built on the CaMeL architecture"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn>=0.34.0,<1.0.0",
    "websockets>=14.0,<16.0",
    "httpx>=0.28.0,<1.0.0",
    "pyyaml>=6.0,<7.0",
    "parso>=0.8",
    "json-repair>=0.30",
    "pydantic>=2.10.0,<3.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "python-json-logger>=3.0.0,<4.0.0",
    "transformers>=4.47.0,<5.0.0",
    "anthropic>=0.42.0,<1.0.0",
    "semgrep>=1.68.0,<2.0.0",
    "slowapi>=0.1.9,<1.0.0",
    "sse-starlette>=2.0.0,<3.0.0",
    "croniter>=3.0.0,<4.0.0",
    "lizard>=1.17.0,<2.0.0",
    "caldav>=1.3,<2.0",
    "radicale>=3.2,<4.0",
    "python-telegram-bot>=21.0,<22.0",
    "aiosmtplib>=3.0,<4.0",
    "asyncpg>=0.30.0,<1.0.0",
    "flashrank>=0.2,<1.0",
    "PyJWT>=2.8,<3.0",
    "cryptography>=42.0,<44.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.25.0,<1.0.0",
    "freezegun>=1.0.0,<2.0.0",
]
worker = [
    # Web & API
    "flask>=3.0,<4.0",
    "beautifulsoup4>=4.12,<5.0",
    "lxml>=5.0,<6.0",
    "aiohttp>=3.9,<4.0",
    "aiofiles>=24.0,<25.0",
    # Data & Files
    "pandas>=2.2,<3.0",
    "openpyxl>=3.1,<4.0",
    "tabulate>=0.9,<1.0",
    "python-docx>=1.1,<2.0",
    "markdown>=3.5,<4.0",
    # Database
    "sqlalchemy>=2.0,<3.0",
    # Code quality
    "black>=24.0",
    "ruff>=0.5",
    "coverage>=7.0,<8.0",
    "hypothesis>=6.100,<7.0",
    "mypy>=1.10,<2.0",
    # Security
    "bcrypt>=4.0,<5.0",
    "passlib>=1.7,<2.0",
    # Image
    "pillow>=10.0,<11.0",
    # System
    "psutil>=6.0,<7.0",
    "watchdog>=4.0,<5.0",
    # CVE auditing (used in monthly rebuild gate)
    "pip-audit>=2.7,<3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "capability: Capability deployment and error handling tests",
    "integration: Requires running services",
    "slow: Takes more than 5 seconds",
]

[tool.setuptools.packages.find]
include = ["sentinel*"]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
